Avoid redrawing all the screen and remove timeout when title is displayed
The title scene was using a timeout to display the letters in the title one by one,
and adding a random delay to simulate a human typing. That timeout was never stoped,
and continued running while the activity was used.

Also to add every letter, all the background image was updated, and that is slow
in the xo-1.75.

This patch stop the timeout when is not needed anymore and update only the title text
in the presentation screen.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>